# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Coda External API",
    "version": "1.0.0",
    "summary": "API for Spaces, screenplay authoring, and project-scoped source breakdown data.",
    "description": "Breakdown routes accept project-bound API keys and MCP tokens. Space and screenplay routes require a signed-in browser session and do not accept project-scoped bearer credentials. Mutating session-authenticated requests also require the CSRF cookie and matching X-Coda-CSRF header. This document intentionally excludes setup, account, session administration, instance administration, membership, role, invitation, ownership-transfer, workspace-layout, project import, trash, and purge operations.",
    "license": { "name": "MIT", "identifier": "MIT" }
  },
  "servers": [{ "url": "/", "description": "The Coda instance that issued the credential." }],
...
  "paths": {
...
    "/api/v1/trackers/{trackerId}": {
      "get": {
        "operationId": "getTracker",
        "summary": "Get a tracker",
        "tags": ["Trackers"],
        "security": [{ "sessionCookie": [] }],
        "parameters": [{ "$ref": "#/components/parameters/TrackerId" }],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/DataEnvelope" },
                    {
                      "type": "object",
                      "properties": { "data": { "$ref": "#/components/schemas/Tracker" } }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": {
            "description": "Request body exceeds the configured transport limit.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" },
          "503": {
            "description": "Request parsing or a required dependency is temporarily unavailable.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "507": {
            "description": "The owner screenplay quota is exhausted.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateTracker",
        "summary": "Rename or describe a tracker",
        "tags": ["Trackers"],
        "security": [{ "sessionCookie": [], "csrfCookie": [], "csrfHeader": [] }],
        "parameters": [{ "$ref": "#/components/parameters/TrackerId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UpdateTrackerInput" } }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/DataEnvelope" },
                    {
                      "type": "object",
                      "properties": { "data": { "$ref": "#/components/schemas/Tracker" } }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": {
            "description": "Request body exceeds the configured transport limit.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" },
          "503": {
            "description": "Request parsing or a required dependency is temporarily unavailable.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "507": {
            "description": "The owner screenplay quota is exhausted.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          }
        }
      }
    },
    "/api/v1/trackers/{trackerId}/fields": {
      "get": {
        "operationId": "listTrackerFields",
        "summary": "List tracker fields in manual order",
        "tags": ["Trackers"],
        "security": [{ "sessionCookie": [] }],
        "parameters": [{ "$ref": "#/components/parameters/TrackerId" }],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/DataEnvelope" },
                    {
                      "type": "object",
                      "properties": { "data": { "$ref": "#/components/schemas/TrackerFieldList" } }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": {
            "description": "Request body exceeds the configured transport limit.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" },
          "503": {
            "description": "Request parsing or a required dependency is temporarily unavailable.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "507": {
            "description": "The owner screenplay quota is exhausted.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createTrackerField",
        "summary": "Create a tracker field",
        "tags": ["Trackers"],
        "security": [{ "sessionCookie": [], "csrfCookie": [], "csrfHeader": [] }],
        "parameters": [{ "$ref": "#/components/parameters/TrackerId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTrackerFieldInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/DataEnvelope" },
                    {
                      "type": "object",
                      "properties": { "data": { "$ref": "#/components/schemas/TrackerField" } }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": {
            "description": "Request body exceeds the configured transport limit.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" },
          "503": {
            "description": "Request parsing or a required dependency is temporarily unavailable.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "507": {
            "description": "The owner screenplay quota is exhausted.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          }
        }
      }
    },
    "/api/v1/trackers/{trackerId}/fields/{fieldId}": {
      "get": {
        "operationId": "getTrackerField",
        "summary": "Get a tracker field",
        "tags": ["Trackers"],
        "security": [{ "sessionCookie": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/TrackerId" },
          { "$ref": "#/components/parameters/FieldId" }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/DataEnvelope" },
                    {
                      "type": "object",
                      "properties": { "data": { "$ref": "#/components/schemas/TrackerField" } }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": {
            "description": "Request body exceeds the configured transport limit.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" },
          "503": {
            "description": "Request parsing or a required dependency is temporarily unavailable.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "507": {
            "description": "The owner screenplay quota is exhausted.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateTrackerField",
        "summary": "Update a tracker field",
        "tags": ["Trackers"],
        "security": [{ "sessionCookie": [], "csrfCookie": [], "csrfHeader": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/TrackerId" },
          { "$ref": "#/components/parameters/FieldId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateTrackerFieldInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/DataEnvelope" },
                    {
                      "type": "object",
                      "properties": { "data": { "$ref": "#/components/schemas/TrackerField" } }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": {
            "description": "Request body exceeds the configured transport limit.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" },
          "503": {
            "description": "Request parsing or a required dependency is temporarily unavailable.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "507": {
            "description": "The owner screenplay quota is exhausted.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "archiveTrackerField",
        "summary": "Archive a tracker field into trash",
        "tags": ["Trackers"],
        "security": [{ "sessionCookie": [], "csrfCookie": [], "csrfHeader": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/TrackerId" },
          { "$ref": "#/components/parameters/FieldId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ArchiveTrackerFieldInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Soft-deletes the field with an optimistic version guard; its key stays reserved.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/DataEnvelope" },
                    {
                      "type": "object",
                      "properties": { "data": { "$ref": "#/components/schemas/ArchiveResult" } }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": {
            "description": "Request body exceeds the configured transport limit.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" },
          "503": {
            "description": "Request parsing or a required dependency is temporarily unavailable.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "507": {
            "description": "The owner screenplay quota is exhausted.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          }
        }
      }
    },
    "/api/v1/trackers/{trackerId}/fields/{fieldId}/reorder": {
      "patch": {
        "operationId": "reorderTrackerField",
        "summary": "Move a tracker field between siblings",
        "tags": ["Trackers"],
        "security": [{ "sessionCookie": [], "csrfCookie": [], "csrfHeader": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/TrackerId" },
          { "$ref": "#/components/parameters/FieldId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ReorderTrackerFieldInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/DataEnvelope" },
                    {
                      "type": "object",
                      "properties": { "data": { "$ref": "#/components/schemas/TrackerField" } }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": {
            "description": "Request body exceeds the configured transport limit.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" },
          "503": {
            "description": "Request parsing or a required dependency is temporarily unavailable.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "507": {
            "description": "The owner screenplay quota is exhausted.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          }
        }
      }
    },
    "/api/v1/trackers/{trackerId}/fields/{fieldId}/options": {
      "post": {
        "operationId": "createTrackerFieldOption",
        "summary": "Add an option to an enum or multi-enum field",
        "tags": ["Trackers"],
        "security": [{ "sessionCookie": [], "csrfCookie": [], "csrfHeader": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/TrackerId" },
          { "$ref": "#/components/parameters/FieldId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTrackerFieldOptionInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/DataEnvelope" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/TrackerFieldOption" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": {
            "description": "Request body exceeds the configured transport limit.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" },
          "503": {
            "description": "Request parsing or a required dependency is temporarily unavailable.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "507": {
            "description": "The owner screenplay quota is exhausted.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          }
        }
      }
    },
    "/api/v1/trackers/{trackerId}/fields/{fieldId}/options/{optionId}": {
      "patch": {
        "operationId": "updateTrackerFieldOption",
        "summary": "Rename or recolor an archived-or-active option",
        "tags": ["Trackers"],
        "security": [{ "sessionCookie": [], "csrfCookie": [], "csrfHeader": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/TrackerId" },
          { "$ref": "#/components/parameters/FieldId" },
          { "$ref": "#/components/parameters/OptionId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateTrackerFieldOptionInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/DataEnvelope" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/TrackerFieldOption" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": {
            "description": "Request body exceeds the configured transport limit.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" },
          "503": {
            "description": "Request parsing or a required dependency is temporarily unavailable.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "507": {
            "description": "The owner screenplay quota is exhausted.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "archiveTrackerFieldOption",
        "summary": "Archive an option (existing values are kept)",
        "tags": ["Trackers"],
        "security": [{ "sessionCookie": [], "csrfCookie": [], "csrfHeader": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/TrackerId" },
          { "$ref": "#/components/parameters/FieldId" },
          { "$ref": "#/components/parameters/OptionId" }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/DataEnvelope" },
                    {
                      "type": "object",
                      "properties": { "data": { "$ref": "#/components/schemas/ArchiveResult" } }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": {
            "description": "Request body exceeds the configured transport limit.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" },
          "503": {
            "description": "Request parsing or a required dependency is temporarily unavailable.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "507": {
            "description": "The owner screenplay quota is exhausted.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Coda API key or MCP token",
        "description": "Use an API key by default. When using an MCP token, also send `X-Coda-Token-Audience: mcp`."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "coda_session",
        "description": "Browser session cookie. Screenplay routes do not accept project-scoped bearer credentials."
      },
      "csrfCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "coda_csrf",
        "description": "CSRF cookie required for mutating session-authenticated requests."
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Coda-CSRF",
        "description": "Must exactly match the coda_csrf cookie."
      }
    },
    "parameters": {
...
      "TrackerId": {
        "name": "trackerId",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request.",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetails" }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid session or bearer credential, including revoked, expired, or wrong-audience bearer credentials.",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetails" }
          }
        }
      },
      "Forbidden": {
        "description": "The credential lacks the required permission, or a session-authenticated mutation failed CSRF validation.",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetails" }
          }
        }
      },
      "NotFound": {
        "description": "The resource does not exist or is not accessible to the caller.",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetails" }
          }
        }
      },
      "Conflict": {
        "description": "The resource version is stale or a domain invariant would be violated.",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetails" }
          }
        }
      },
      "TooManyRequests": {
        "description": "Request rate limit exceeded.",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetails" }
          }
        }
      },
...
    },
    "schemas": {
...
      "DataEnvelope": {
        "type": "object",
        "required": ["data"],
        "properties": {
          "data": {},
          "meta": {
            "type": "object",
            "description": "Pagination or response metadata when present.",
            "additionalProperties": true,
            "properties": { "nextCursor": { "type": ["string", "null"] } }
          }
        }
      },
...
      "Tracker": {
        "allOf": [
          { "$ref": "#/components/schemas/TrackerSummary" },
          {
            "type": "object",
            "required": ["access"],
            "properties": {
              "access": {
                "type": "object",
                "required": ["permissions"],
                "properties": { "permissions": { "type": "array", "items": { "type": "string" } } }
              }
            }
          }
        ]
      },
...
      "TrackerFieldList": {
        "type": "array",
        "items": { "$ref": "#/components/schemas/TrackerField" }
      },
...
    }
  },
...
}